+++ changelog.md
@@ -4,10 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [planned]
+## [v0.2.1]
 ### Changed
+- change and other actions not default behaviour
+- Warning message if no credentials found in config or credentials
 
-## [Unreleased]
+## [v0.1.0]
 ### Added
 - do not change dir, when current pwd is subdirectory
 - Change "workdir" entry